Update the test.
This commit updates the test so that both the middleware and proxy share the
`outgoing-handler` from the host.

The proxy now sends a request via the host's `outgoing-handler` first to verify
that it doesn't get modified, then sends another request through the
middleware, which modifies the path and sends the request through the same
`outgoing-handler`.